feat(lib): convert distressing content question to typescript
